+++ src/pages/Dashboard/Automations/AddAutomationCard.vue
@@ -71,7 +71,8 @@ export default {
       rules: {
         required: val => !!val || 'Required',
         notNull: val => val > 0 || 'Duration must be greater than 0 seconds'
-      }
+      },
+      animated: false
     }
   },
   computed: {
@@ -288,6 +289,14 @@ export default {
     actionIcon(type) {
       return actionTypes.find(a => a.actionType == type)?.icon
     },
+    addHint() {
+      clearTimeout(this.animationTimeout)
+      this.animated = true
+
+      this.animationTimeout = setTimeout(() => {
+        this.animated = false
+      }, 1500)
+    },
     buttonColor(selectedStep, otherStep) {
       // const stepComplete = this.steps[selectedStep]
       // const otherComplete = this.steps[otherStep]
@@ -323,6 +332,12 @@ export default {
       return this[`stateGroup${group}`]
     },
     switchStep(selectedStep) {
+      if (this.step.name == selectedStep) {
+        this.addHint()
+      } else {
+        clearTimeout(this.animationTimeout)
+      }
+
       this.previousStep = this.step.name
 
       if (this.step.name === 'openDuration')
@@ -727,6 +742,19 @@ export default {
           </v-btn>
 
           <v-btn
+            v-else-if="agentOrFlow === 'this'"
+            :style="{ 'text-transform': 'none', 'min-width': '0px' }"
+            color="accentPink"
+            class="px-0 pb-1 ml-1 text-h5 d-inline-block text-truncate"
+            text
+            :disabled="addAction"
+            max-width="500px"
+            @click="addHint"
+          >
+            <span>this</span>
+          </v-btn>
+
+          <v-btn
             v-else
             :style="{ 'text-transform': 'none', 'min-width': '0px' }"
             :color="buttonColor('openAgentOrFlow')"
@@ -852,7 +880,7 @@ export default {
       <!-- OPEN AGENT OR FLOW -->
       <div v-if="step.name === 'openAgentOrFlow'" key="openAgentOrFlow">
         <div>
-          <div class="mb-2 text-subtitle-1 font-weight-light">
+          <div class="mb-2 text-subtitle-1 accentPink--text font-weight-light">
             Choose an automation:
           </div>
           <v-row>
@@ -866,6 +894,7 @@ export default {
                   agentOrFlow === item.type ? 'accentPink' : 'utilGrayMid'
                 "
                 class="mr-4 cursor-pointer text-h6 font-weight-light remove--disabled"
+                :class="{ flash: animated }"
                 :disabled="!hasPermission(item.permission)"
                 :input-value="agentOrFlow === item.type"
                 @click="selectAgentOrFlow(item.type)"
@@ -922,6 +951,7 @@ export default {
                 small
                 elevation="0"
                 color="primary"
+                :class="{ flash: animated }"
                 @click="showAgentConfigForm = true"
               >
                 <v-icon small class="mr-2">fa-plus</v-icon>
@@ -937,7 +967,8 @@ export default {
                   :class="{
                     active:
                       selectedAgentConfig &&
-                      selectedAgentConfig.id === config.id
+                      selectedAgentConfig.id === config.id,
+                    flash: animated
                   }"
                   @click="selectAgentConfig(config)"
                 >
@@ -1029,7 +1060,7 @@ export default {
                   <div
                     v-ripple
                     class="chip-bigger d-flex align-center justify-start pa-2 cursor-pointer"
-                    :class="{ active: includesFlow(item) }"
+                    :class="{ active: includesFlow(item), flash: animated }"
                     @click="selectFlow(item)"
                   >
                     <div style="width: auto;" class="text-body-1 text-truncate">
@@ -1086,6 +1117,7 @@ export default {
                     ? 'accentPink'
                     : 'utilGrayMid'
                 "
+                :class="{ flash: animated }"
                 class="mr-4 cursor-pointer text-h6 font-weight-light remove--disabled"
                 :input-value="flowEventType.name === item.name"
                 :disabled="!hasPermission(item.permission)"
@@ -1136,8 +1168,10 @@ export default {
               v-model="seconds"
               type="number"
               :rules="[rules.required, rules.notNull]"
+              :class="{ flash: animated }"
               persistent-hint
               outlined
+              hide-details
               @keydown.enter="closeSeconds"
             ></v-text-field>
           </v-col>
@@ -1182,7 +1216,7 @@ export default {
               : `Select ${item} and all connected states`
           "
           class="mr-2"
-          :class="{ 'white--text': dynamicStateGroup(item) }"
+          :class="{ 'white--text': dynamicStateGroup(item), flash: animated }"
           @click="selectStateGroup(item)"
         >
           {{ item }}
@@ -1194,7 +1228,7 @@ export default {
             :key="item"
             v-ripple
             class="d-inline-block chip-small pa-2 mr-4 my-2 cursor-pointer text-body-1"
-            :class="{ active: chosenStates.includes(item) }"
+            :class="{ active: chosenStates.includes(item), flash: animated }"
             @click="selectStates(item)"
           >
             {{ item }}
@@ -1272,7 +1306,8 @@ export default {
               class="chip-small pa-2 mr-4 my-2 cursor-pointer text-body-1 d-inline-block"
               :class="{
                 active: chosenAction && chosenAction.id === item.id,
-                disabled: !hasPermission('feature:api-action')
+                disabled: !hasPermission('feature:api-action'),
+                flash: animated
               }"
               @click="selectAction(item)"
             >
@@ -1295,7 +1330,10 @@ export default {
               :key="item.id"
               v-ripple
               class="chip-small pa-2 mr-4 my-2 cursor-pointer text-body-1 d-inline-block"
-              :class="{ active: chosenAction && chosenAction.id === item.id }"
+              :class="{
+                active: chosenAction && chosenAction.id === item.id,
+                flash: animated
+              }"
               @click="selectAction(item)"
             >
               <v-icon small class="mr-2">{{
@@ -1355,6 +1393,7 @@ export default {
 
 <style lang="scss" scoped>
 .chip-bigger {
+  border-radius: 5px;
   height: 60px;
   position: relative;
   transition: all 50ms;
@@ -1385,6 +1424,7 @@ export default {
 }
 
 .chip-small {
+  border-radius: 5px;
   max-width: fit-content;
   position: relative;
   transition: all 50ms;
@@ -1428,6 +1468,27 @@ export default {
       position: absolute;
       width: 100%;
     }
+  }
+}
+
+.flash {
+  &:not(:disabled):not(.disabled) {
+    animation: flash 1.5s;
+  }
+}
+
+@keyframes flash {
+  0%,
+  50%,
+  100% {
+    background-color: transparent;
+    border-color: currentColor !important;
+  }
+
+  25%,
+  75% {
+    background-color: rgba(59, 141, 255, 0.2);
+    border-color: var(--v-primary-base) !important;
   }
 }
 </style>