+++ Changelog.md
@@ -4,7 +4,9 @@
 
 ### Features and Improvements
 
+- Adds animated hint for users creating new automations - [#784](https://github.com/PrefectHQ/ui/pull/784)
 - Adds a loading indicator for inital page load - [#789](https://github.com/PrefectHQ/ui/pull/789)
+- Take usage tile off dashboard and add to account page - [#782](https://github.com/PrefectHQ/ui/pull/782)
 
 ### Bugfixes